Add testing code for new  external texture

+++ shell/platform/android/io/flutter/embedding/engine/renderer/FlutterRenderer.java
@@ -120,7 +120,7 @@ public class FlutterRenderer implements TextureRegistry {
   }
 
   @Override
-  public void onShareFrameAvaliable(int textureIndex) {
+  public void onShareFrameAvaliable(long textureIndex) {
     markTextureFrameAvailable(textureIndex);
   }
